Drastically simplify scope-increase
Turn it into a token, so we get matching of the newline and match failure for
free. Use a look-ahead assertion for finding the indentation level. It only
matches if the current level is matched and some more whitespace, so there's no
need to check whether the new indentation level is larger than the current one.
The rule would fail to match otherwise anyway. Just take a note of the current
position in the look-ahead assertion to calculate the new indentation level.