Get rid of the extra reference around each object
Lvalue subs let us assign to variables and we convert all values into
objects anyway, so there's no need to push around references to objects
(which themselves are again blessed references).

Sadly do { } blocks cannot return lvalues, so we had to move throwing
NameErrors into the getvar sub (or rather StackFrame's __getattr__). But
this greatly reduces the size of the generated code, so in the end it's
a good change.