DTML: fix attribute confusion
Value attributes can be just numbers or strings but also words that we have to
resolve. In fact the old compiler did not look too closely at this kind
of attribute, so most values were just funneled through into the Perl
code unmodified. This means that one could execute arbitrary Perl code
as long as it happened to look like Python code. That's clearly insane
and unsafe, so we draw the line at numbers, strings and words.