Fix assigments to list/dict subscripts
Getting rid of the extra refs made assignment a bit harder. We got
around this using lvalue subs for resolving names and we have to do the
same for subscripts. This means that throwing the exception on unknown
keys also has to be done by __getitem__ as do blocks can not be lvalues.