Fix Python2::Type objects escaping into Perl arrays
The iteration variable of a Perl for loop is aliased to the iterator
value. This means that assignments do not only affect this variable, but
change the list we're iterating over. Thus assigning the converted value
actually changed the array. Fix by using different variables.

As a bonus this improves sort performance by only converting every value
once.